Export limit values seem to be scaled. Add scaling factor.

+++ solax/integrationpluginsolax.cpp
@@ -2023,7 +2023,7 @@ void IntegrationPluginSolax::executeAction(ThingActionInfo *info)
             thing->setStateValue(solaxX3InverterTCPEnableExportLimitStateTypeId, enableExportLimit);
             const auto exportLimit = thing->stateValue(solaxX3InverterTCPExportLimitStateTypeId).toDouble();
             const auto ratedPower = thing->stateValue(solaxX3InverterTCPNominalPowerStateTypeId).toDouble();
-            limitToSet = static_cast<quint16>(enableExportLimit ? exportLimit : ratedPower);
+            limitToSet = enableExportLimit ? exportLimit : ratedPower;
 
         } else if (action.actionTypeId() == solaxX3InverterTCPExportLimitActionTypeId) {
             const auto exportLimit =
@@ -2031,7 +2031,7 @@ void IntegrationPluginSolax::executeAction(ThingActionInfo *info)
             thing->setStateValue(solaxX3InverterTCPExportLimitStateTypeId, exportLimit);
             const auto exportLimitEnabled = thing->stateValue(solaxX3InverterTCPEnableExportLimitStateTypeId).toBool();
             if (exportLimitEnabled) {
-                limitToSet = static_cast<quint16>(exportLimit);
+                limitToSet = exportLimit;
             }
         } else {
             Q_ASSERT_X(false, "executeAction", QString("Unhandled action: %1").arg(actionType.name()).toUtf8());
@@ -2040,7 +2040,7 @@ void IntegrationPluginSolax::executeAction(ThingActionInfo *info)
         }
 
         if (limitToSet.isValid()) {
-            const auto limit = static_cast<quint16>(limitToSet.toUInt());
+            const auto limit = limitToSet.toFloat();
             qCInfo(dcSolax()) << "\tSetting power limit:" << limit;
             const auto reply = connection->setWriteExportLimit(limit);
             connect(reply, &QModbusReply::finished, reply, &QModbusReply::deleteLater);
@@ -2077,14 +2077,14 @@ void IntegrationPluginSolax::executeAction(ThingActionInfo *info)
             thing->setStateValue(solaxX3InverterRTUEnableExportLimitStateTypeId, enableExportLimit);
             const auto exportLimit = thing->stateValue(solaxX3InverterRTUExportLimitStateTypeId).toDouble();
             const auto ratedPower = thing->stateValue(solaxX3InverterRTUNominalPowerStateTypeId).toDouble();
-            limitToSet = static_cast<quint16>(enableExportLimit ? exportLimit : ratedPower);
+            limitToSet = enableExportLimit ? exportLimit : ratedPower;
         } else if (action.actionTypeId() == solaxX3InverterRTUExportLimitActionTypeId) {
             const auto exportLimit =
                     action.paramValue(solaxX3InverterRTUExportLimitActionExportLimitParamTypeId).toDouble();
             thing->setStateValue(solaxX3InverterRTUExportLimitStateTypeId, exportLimit);
             const auto exportLimitEnabled = thing->stateValue(solaxX3InverterRTUEnableExportLimitStateTypeId).toBool();
             if (exportLimitEnabled) {
-                limitToSet = static_cast<quint16>(exportLimit);
+                limitToSet = exportLimit;
             }
         } else {
             Q_ASSERT_X(false, "executeAction", QString("Unhandled action: %1").arg(actionType.name()).toUtf8());
@@ -2093,7 +2093,7 @@ void IntegrationPluginSolax::executeAction(ThingActionInfo *info)
         }
 
         if (limitToSet.isValid()) {
-            const auto limit = static_cast<quint16>(limitToSet.toUInt());
+            const auto limit = limitToSet.toFloat();
             qCInfo(dcSolax()) << "\tSetting power limit:" << limit;
             const auto reply = connection->setWriteExportLimit(limit);
             connect(reply, &ModbusRtuReply::finished, reply, &ModbusRtuReply::deleteLater);
